Safari-611.1.10/LayoutTests/fast/canvas/webgl/vertexAttribPointer-with-bad-offset-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_VALUE: vertexAttribPointer: bad offset
PASS: vertexAttribPointer should have an error.
